move tests to test_set_ops.py
