+++ pandas/tests/indexing/multiindex/test_multiindex.py
@@ -407,17 +407,6 @@ class TestMultiIndexBasic(object):
             with catch_warnings(record=True):
                 df.ix[name, 'new_col'] = new_vals
 
-    def test_multiindex_symmetric_difference(self):
-        # GH 13490
-        idx = MultiIndex.from_product([['a', 'b'], ['A', 'B']],
-                                      names=['a', 'b'])
-        result = idx ^ idx
-        assert result.names == idx.names
-
-        idx2 = idx.copy().rename(['A', 'B'])
-        result = idx ^ idx2
-        assert result.names == [None, None]
-
     def test_multiindex_contains_dropped(self):
         # GH 19027
         # test that dropped MultiIndex levels are not in the MultiIndex
@@ -782,30 +771,6 @@ class TestMultiIndexBasic(object):
         expected = df['routine1', 'result1', '']
         expected = expected.rename(('routine1', 'result1'))
         tm.assert_series_equal(result, expected)
-
-    def test_mixed_depth_insert(self):
-        arrays = [['a', 'top', 'top', 'routine1', 'routine1', 'routine2'],
-                  ['', 'OD', 'OD', 'result1', 'result2', 'result1'],
-                  ['', 'wx', 'wy', '', '', '']]
-
-        tuples = sorted(zip(*arrays))
-        index = MultiIndex.from_tuples(tuples)
-        df = DataFrame(randn(4, 6), columns=index)
-
-        result = df.copy()
-        expected = df.copy()
-        result['b'] = [1, 2, 3, 4]
-        expected['b', '', ''] = [1, 2, 3, 4]
-        tm.assert_frame_equal(result, expected)
-
-    def test_dataframe_insert_column_all_na(self):
-        # GH #1534
-        mix = MultiIndex.from_tuples([('1a', '2a'), ('1a', '2b'), ('1a', '2c')
-                                      ])
-        df = DataFrame([[1, 2], [3, 4], [5, 6]], index=mix)
-        s = Series({(1, 1): 1, (1, 2): 2})
-        df['new'] = s
-        assert df['new'].isna().all()
 
     def test_set_column_scalar_with_ix(self, multiindex_dataframe_random_data):
         frame = multiindex_dataframe_random_data
